linting and re-organizing
